+++ packages/perseus/src/widgets/iframe/iframe.tsx
@@ -8,7 +8,6 @@
  */
 
 import getIFramePublicWidgetOptions from "@khanacademy/perseus-core/src/widgets/iframe/iframe-util";
-import {scoreIframe} from "@khanacademy/perseus-score";
 import $ from "jquery";
 import * as React from "react";
 import _ from "underscore";
@@ -171,8 +170,5 @@ export default {
     widget: Iframe,
     // Let's not expose it to all content creators yet
     hidden: true,
-    // TODO(LEMS-2656): remove TS suppression
-    // @ts-expect-error: Type 'UserInput' is not assignable to type 'PerseusIframeUserInput'.
-    scorer: scoreIframe,
     getPublicWidgetOptions: getIFramePublicWidgetOptions,
 } satisfies WidgetExports<typeof Iframe>;